restructure database scripts
